A.4: RAII owning-handle sweep + 2 input-hardening edits (Ave robustus!)
Memory/resource/arithmetic phase. Scoping audit (RAII + arithmetic
overflow over base/src) found no live memory-safety bug — A.3 had
already hardened the arithmetic surface. Three fragility/defense-in-depth
items closed, none exploitable:

- RAII (the only owning-raw-pointer gap): IO::ZipReader + IO::ZipWriter
  now hold the libzip handle in std::unique_ptr<zip_t, decltype(&zip_close)>
  instead of a bare zip* released by a hand-written closeArchive(). The
  deleter runs on every path; the classes become move-only. New
  test_ZipArchive.cpp (round-trip + reader/writer abandoned-without-close,
  destructor release proven leak-free under ASan). Doctrine added to
  docs/error-handling.md §5.2.
- FileFormatMDx.hpp:324 (MDL skin read): read the already-validated
  uint64 skinSize instead of recomputing the size from signed dims, so the
  read size can no longer drift from the resized buffer. Verified zero
  behavioral delta (divergent cases rejected by exceedsStream first);
  clarity edit + inline comment, no distinguishing test possible.
- FileFormatOBJ.hpp:1393 (resolveIndex negative indices): widened the
  int32_t cast of listSize to int64_t (UB once a list exceeds INT_MAX);
  behavior identical for realistic inputs. New regression test
  negativeRelativeIndicesResolveLikePositive.

Suite 1751/1751 green in Release and under ASan/UBSan (1741->1751:
+3 Zip, +1 OBJ); projet-alpha cascade compiles and links.

+++ src/VertexFactory/FileFormatOBJ.hpp
@@ -1390,7 +1390,7 @@ namespace EmEn::Base::VertexFactory
 			{
 				if ( index < 0 )
 				{
-					return static_cast< index_data_t >(static_cast< int32_t >(listSize) + index + 1);
+					return static_cast< index_data_t >(static_cast< int64_t >(listSize) + index + 1);
 				}
 
 				return static_cast< index_data_t >(index);
